Fixes and improvements:
- Fix: Verbosity flag can now be given several times
- New flag: -e/--err-all makes the interpreter ignore catch marks and
always error out
- Logs are now printed in red
- Expectable errors should now raise a CodeException that can be caught
with a set catch mark (see below)
- New commands: NSEW (north, south, east, west) acts like ^v><, but
afterwards turns clockwise (north becomes east, etc.)
- New command: @ sets a catch mark: When a CodeException is raised
anywhere, instead of crashing the IP is moved to the last seen catch
mark. This might change in the future to be a stack of catch marks.
- New command: ? moves in a random direction.
- New command: R pushes a random number on the stack (between 0 and 1)
- New commands: P/e push π/e on the stack.
- New command: ~ is a unary invert. Does a negation for numbers, and
currently crashes for strings.
- New command: O goes to the origin (0,0 normally, or the bottom right
corner if the direction is reversed).
- New command: , gets a single character without waiting for a newline.
^C and ^Z are specially handled and reraised.
- New command: . prints the first character of the top stack element. If
it's a number, it transforms it into a character instead (chr).
- New command: & raises a generic CodeException (to be caught by a catch
mark).

This version makes Aceto incompatible with Windows. This may stay that
way.